requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
DateTime==4.4
google-auth==2.6.6
google-auth-oauthlib==0.5.1
greenlet==1.1.2
gspread==5.3.2
idna==3.3
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
xyaml==1.0.2
zope.interface==5.4.0
